extract errors and read helpers into separate files
